// src/routes/type.ts
export type PowerUp = 
{
    itemID: number;
    itemTypeID: number;
    vcoinPrice: number;
    isPremium: boolean;
    filePath: string;
    isActive: boolean;
    name: string;
    description: string;
}
// ...
export type UserPowerUp = 
{
    itemId: number;
    userPlayerId: number;
    quantity: number;
    filePath: string;
    name: string;
    description: string;
}